Fix invalid numbers in lists not treated as errors
For options like -p, -P, -a, -A, and -x where the elements were parsed
into doubles, parse errors were added to the error list, but the state
of the ProgramOptions object was not set to invalid. This was fixed by
setting the state to invalid in that case.

A more permanent fix would be to create a separate member function to
handle both appending an error and changing the state to invalid.