Improve adding errors in ProgramOptions
Because ProgramOptions::valid_ is used to indicate whether the options
are valid, any time an error was encountered, the error had to manually
be pushed into the error list and valid_ had to be set to false. This
led to several bugs where the error would be pushed, but valid_ was not
properly set. Therefore, this was factored into a function
ProgramOptions::add_error, which both adds the error and sets the state
to invalid.