Generate usage message from file
Previously, the usage message was specified using a normal string
literal in `disscalc/output.cpp`. Unfortunately, formatting the contents
of a string literal is not very effective, so I would typically edit the
usage message in a file, then copy it into a string literal in the
source file.

Clearly, it would be much more convenient if that step could be done
automatically. Cmake has several features that would facilitate this:

- `file(READ)` can be used to read the usage message into a variable.
- `configure_file` can be used to configure a source file with the
  contents of that variable (being the contents of the file) inserted.

The problem with these is that all of this is done at configure time
rather than build time, and cmake doesn't know that the file containing
the usage message is a dependency for the target, so cmake has to be
manually reconfigured every time that file changes.

As it turns out, cmake has a feature for this exact thing: the property
`CMAKE_CONFIGURE_DEPENDS`. This contains a list of files that are
dependencies to the configuration of cmake itself. The usage message
file can be specified as a dependency that way, and cmake will
automatically reconfigure whenever it is changed.